refactor: extract SwiftUI components into separate files
- Extract SidebarView into separate file
- Extract NativeSplitView and SplitHostController into separate file
- Extract PreviewPanel and video playback components into separate file
- Extract ControlPanel and SelectionSummary into separate file
- Extract StatusPanel and MetricsGrid into separate file
- Extract SubtitleListPanel and SubtitleListView into separate file
- Extract utility functions (formattedTime, formattedTimestamp, etc.) into separate file
- Keep ContentView as main coordinator file
- Improve code organization and maintainability
